requirements_health_check.txt
requests>=2.28.0
psycopg2-binary>=2.9.0
pika>=1.3.0
